@@ -2,6 +2,7 @@ using System;
 using Microsoft.Data.SqlClient;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using Dapper;
 using HybridDb.Config;
@@ -20,9 +21,9 @@ namespace HybridDb.Tests.Migrations.BuiltIn
         // To export data for use in these tests, use "Tasks > Generate Scripts" in SQL Management Studio.
         // Under "Advanced" select "Schema and data" for "Types of data to script"
 
-        // TODO: Test at et dokument, der loades uden configureret design via migration ikke fors�ges gemt i 
-        // databasen i en forkert tabel (Documents) og derved lader migrations g� i uendelig l�kke. 
-        // M�ske skal ManagedEntity have Design p� sig?
+        // TODO: Test at et dokument, der loades uden configureret design via migration ikke forsøges gemt i 
+        // databasen i en forkert tabel (Documents) og derved lader migrations gå i uendelig løkke. 
+        // Måske skal ManagedEntity have Design på sig?
         public HybridDb_1_x_x_to_2_x_x_Part1_Tests(ITestOutputHelper output) : base(output) { }
 
         [Fact]
